user_auth/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-cors-headers==3.5.0
django-crontab==0.7.1
django-email-verification==0.0.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
Flask==1.1.2
idna==2.10
importlib-metadata==3.1.0
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
asgiref==3.3.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-cors-headers==3.5.0
django-crontab==0.7.1
django-email-verification==0.0.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
Flask==1.1.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
importlib-metadata==3.1.0
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauth2client==4.1.3
openapi-codec==1.3.2
packaging==20.4
Pillow==8.1.0
protobuf==3.15.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.7.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
python-dateutil==2.8.1